Improve language support by using `lang` attributes, add note about non-Chrome webkit browsers not working
Webkit browsers return a truthy value when checking SpeechRecognition support, but they actually will fail when calling start.
There does not seem to be a reliable way of checking browser support without trying to invoke the SpeechRecognition API and looking for the `network` error.
No way to make this check without user intevention of allowing permissions.